Revert "develop-tutorial"
