--- src/main/java/com/kach/tuts/dao/TutorialDao.java
@@ -1,7 +1,0 @@
-package com.kach.tuts.dao;
-
-
-import com.kach.tuts.domain.Tutorial;
-
-public interface TutorialDao extends CrudDao<Tutorial> {
-}
