--- src/main/java/com/kach/tuts/domain/Tutorial.java
@@ -1,38 +1,0 @@
-package com.kach.tuts.domain;
-
-import lombok.*;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Table;
-import java.util.Objects;
-
-@Getter
-@Setter
-@RequiredArgsConstructor
-@NoArgsConstructor
-@Entity
-@Table(name = "tutorial")
-public class Tutorial extends BaseEntity{
-    @NonNull
-    @Column
-    String title;
-
-    @NonNull
-    @Column
-    String description;
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Tutorial that = (Tutorial) o;
-        return title.equals(that.title) &&
-                description.equals(that.description);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(title, description);
-    }
-}
