--- src/main/java/com/kach/tuts/services/TutorialService.java
@@ -1,7 +1,0 @@
-package com.kach.tuts.services;
-
-
-import com.kach.tuts.domain.Tutorial;
-
-public interface TutorialService extends CrudService<Tutorial> {
-}
