+++ src/main/java/com/kach/tuts/services/impl/TutorialServiceImpl.java
@@ -1,45 +1,44 @@
 package com.kach.tuts.services.impl;
 
-
-import com.kach.tuts.dao.TutorialDao;
-import com.kach.tuts.domain.Tutorial;
-import com.kach.tuts.services.TutorialService;
+import com.kach.tuts.dao.TutorialStepDao;
+import com.kach.tuts.domain.TutorialStep;
+import com.kach.tuts.services.TutorialStepService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
-public class TutorialServiceImpl implements TutorialService {
-    private TutorialDao tutorialDao;
+public class TutorialServiceImpl implements TutorialStepService {
+    private TutorialStepDao tutorialStepDao;
 
     @Autowired
-    public void setHashtagDao(TutorialDao tutorialDao) {
-        this.tutorialDao = tutorialDao;
+    public void setTutorialStepDao(TutorialStepDao tutorialStepDao) {
+        this.tutorialStepDao = tutorialStepDao;
     }
 
     @Override
-    public Tutorial save(Tutorial tutorial) {
-        return tutorialDao.save(tutorial);
+    public TutorialStep save(TutorialStep tutorialStep) {
+        return this.tutorialStepDao.save(tutorialStep);
     }
 
     @Override
-    public void update(Tutorial tutorial) {
-        tutorialDao.update(tutorial);
+    public void update(TutorialStep tutorialStep) {
+        this.tutorialStepDao.update(tutorialStep);
     }
 
     @Override
-    public Tutorial getById(Long id) {
-        return tutorialDao.getById(id);
+    public TutorialStep getById(Long id) {
+        return this.tutorialStepDao.getById(id);
     }
 
     @Override
-    public List<Tutorial> getAll() {
-        return tutorialDao.getAll();
+    public List<TutorialStep> getAll() {
+        return this.tutorialStepDao.getAll();
     }
 
     @Override
-    public void delete(Tutorial hashtag) {
-        tutorialDao.delete(hashtag);
+    public void delete(TutorialStep tutorialStep) {
+        this.tutorialStepDao.delete(tutorialStep);
     }
 }
